move prompt user into user_interface_helpers where it belongs
