+++ scripts/withdraw.js
@@ -1,6 +1,8 @@
-const { signAndSend, promptUser } = require("./utils/sign_and_send")(web3, artifacts)
+const { signAndSend } = require("./utils/sign_and_send")(web3, artifacts)
 const { getSafe } = require("./utils/trading_strategy_helpers")(web3, artifacts)
+const { promptUser } = require("./utils/user_interface_helpers")
 const prepareWithdraw = require("./wrapper/withdraw")(web3, artifacts)
+
 
 const argv = require("./utils/default_yargs")
   .option("masterSafe", {
